Allow nested document nodes
This is useful when constructing ASTs that contain a mix of computed and
parsed data. "sub documents" that are parsed have a single root document
node, and rather than having to extract its children, those can be
placed as-is in a larger parse tree.

All I did to implement this is simply remove the restriction. At least
the commonmark renderer worked just fine.